requirements.txt
python-telegram-bot==21.6
aiosqlite==0.21.0
APScheduler==3.10.4
